zlib: compat with 2286d8c17e5bf3d4312357eba9a845262068f007

+++ zlib.rs
@@ -3,14 +3,17 @@ use std;
 import core::ptr;
 import core::vec;
 
+import ctypes::c_int;
+
 export compress, uncompress;
 
 #[link_name = "z"]
 #[abi = "cdecl"]
 native mod _native {
     fn compressBound(len: u64) -> u64;
-    fn compress2(dest: *u8, destlen: *u64, src: *u8, srclen: u64, level: int) -> int;
-    fn uncompress(dest: *u8, destlen: *u64, src: *u8, srclen: u64) -> int;
+    fn compress2(dest: *u8, destlen: *u64, src: *u8, srclen: u64,
+                 level: c_int) -> c_int;
+    fn uncompress(dest: *u8, destlen: *u64, src: *u8, srclen: u64) -> c_int;
 }
 
 fn compress(src: [u8], level: int) -> [u8] unsafe {
@@ -22,10 +25,10 @@ fn compress(src: [u8], level: int) -> [u8] unsafe {
     let pdest = vec::unsafe::to_ptr::<u8>(dest);
     let psrc = vec::unsafe::to_ptr::<u8>(src);
     let pdestlen = ptr::addr_of::<u64>(destlen);
-    let r = _native::compress2(pdest, pdestlen, psrc, srclen, level);
+    let r = _native::compress2(pdest, pdestlen, psrc, srclen, level as c_int);
     // XXX: 0 == Z_OK
-    if r != 0 {
-        fail #fmt["zlib compress2() returned %d", r];
+    if r != 0 as c_int {
+        fail #fmt["zlib compress2() returned %d", r as int];
     }
     vec::slice::<u8>(dest, 0u, destlen as uint)
 }
@@ -38,8 +41,8 @@ fn uncompress(src: [u8], _destlen: u64) -> [u8] unsafe {
     let pdest = vec::unsafe::to_ptr::<u8>(dest);
     let psrc = vec::unsafe::to_ptr::<u8>(src);
     let r = _native::uncompress(pdest, pdestlen, psrc, srclen);
-    if r != 0 {
-        fail #fmt["zlib uncompress() returned %d", r];
+    if r != 0 as c_int {
+        fail #fmt["zlib uncompress() returned %d", r as int];
     }
     vec::slice::<u8>(dest, 0u, destlen as uint)
 }
